+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - adjusted Handheld Dualwield boni to +50% instead of 150% (whooops)
 - fixed issue with producing broken Item when removing last Module
 - fixed crash with Golden Age Combat
+- fixed sweeping not working on forge